run rubocop --auto-correct -A

+++ db/migrate/20201210082826_add_service_name_to_active_storage_blobs.active_storage.rb
@@ -15,8 +15,6 @@ class AddServiceNameToActiveStorageBlobs < ActiveRecord::Migration[6.0]
   end
 
   def down
-    if table_exists?(:active_storage_blobs)
-      remove_column :active_storage_blobs, :service_name
-    end
+    remove_column :active_storage_blobs, :service_name if table_exists?(:active_storage_blobs)
   end
 end
